Refactoring end-to-end tests

+++ sh/test-clean-reports.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
-rm -rf ./results/playwright-report
-rm -f ./results/*
-
+if [[ -d results ]]; then
+  rm -rf ./results/playwright-report
+  rm -f ./results/*
+fi
